% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spicedata-package.R
\docType{package}
\name{spicedata-package}
\alias{spicedata}
\alias{spicedata-package}
\title{spicedata: Data sets of Agrodiversity Experiment}
\description{
Combine all data sets in the agrodiversity experiment.
}
\author{
\strong{Maintainer}: Litty Mathew \email{littymathew1996@gmail.com}

}
\keyword{internal}
