% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studioalbumtracksdoc.R
\docType{data}
\name{studio_album_tracks}
\alias{studio_album_tracks}
\title{studio album data}
\format{
A dataframe with 31 rows and 25 variables:
\describe{
\item{Site}{Site ID number}
\item{Day}{Day of data collected}
}
}
\source{
\url{https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/j.1365-2745.2007.01225.x}
}
\usage{
studio_album_tracks
}
\description{
got from github
}
\keyword{datasets}
