% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_album_tracks.R
\docType{data}
\name{studio_album_tracks}
\alias{studio_album_tracks}
\title{Spice Girls Studio Album Tracks}
\format{
A tibble with 31 rows and 25 variables:
\describe{
\item{artist_name}{chr Name of the artist}
\item{artist_id}{chr Unique ID number of artist}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2021/2021-12-14}
}
\usage{
studio_album_tracks
}
\description{
Audio features of each song from the three studio albums by the Spice Girls. From Spotify.
}
\keyword{datasets}
