% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio_album_tracks.R
\docType{data}
\name{studio_album_tracks}
\alias{studio_album_tracks}
\title{Spice data}
\format{
A tibble with 31 rows and 25 variables:
\describe{
\item{artist_name}{chr Name}
\item{artist_id}{chr Month price was recorded. Ranges from 1-12 for January - December}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2021/2021-12-14}
}
\usage{
studio_album_tracks
}
\description{
Shrimp are classified by size, 0-15 shrimp per pound, 15-20 shrimp per pound,
etc. A smaller number per pound indicates larger shrimp. Nominal prices are
total monthly value of brown shrimp andings within size class divided by
total monthly landings within the size class.
}
\keyword{datasets}
